* Support linking to a specific line in a note on the render page * Split out stamping logic into its won library * fix bug in definition patttern
